Add files via upload
Objective:
The main objective of this project is to design and implement a robust data preprocessing system that addresses common challenges such as missing values, outliers, inconsistent formatting, and noise. By performing effective data preprocessing, the project aims to enhance the quality, reliability, and usefulness of the data for machine learning.
Explored the data, found the unique values in each feature and  its length.
Performed the statistical analysis and renamed  the columns. Found the missing and inappropriate values, treated them appropriately.
Removed all duplicate rows. Found the outliers. Replaced the value 0 in age as NaN. Treated the null values in all columns using any measures(removing/ replace the values with mean/median/mode).Filtered the data with age >40 and salary<5000. A scatterplot  chart is created with age and salary. Counted the number of people from each place and represented it visually. Converted the categorical variables into numerical representations using techniques such as one-hot encoding, label encoding, making them suitable for analysis by machine learning algorithms. After the process of encoding, performed the scaling of the features using standardscaler and minmaxscaler.